Move app_modules/ to lib/
That's the standard directory for, well, libs.
